again, crash in p4 descend; stop ctd read

+++ MPC/CTD/CTD.c
@@ -162,13 +162,13 @@ bool CTD_GetPrompt() {
   memset(stringin, 0, STRING_SIZE);
   TURxFlush(devicePort);
   TUTxPutByte(devicePort, '\r', true);
-  ch=TURxGetByteWithTimeout(devicePort, 10); //cr echo
-  ch=TURxGetByteWithTimeout(devicePort, 10); //lf echo
+  ch=TURxGetByteWithTimeout(devicePort, (short) 10); //cr echo
+  ch=TURxGetByteWithTimeout(devicePort, (short) 10); //lf echo
   GetStringWait(stringin, (short) 2000);
   TURxFlush(devicePort);
   TUTxPutByte(devicePort, '\r', true);
-  ch=TURxGetByteWithTimeout(devicePort, 10);
-  ch=TURxGetByteWithTimeout(devicePort, 10);
+  ch=TURxGetByteWithTimeout(devicePort, (short) 10);
+  ch=TURxGetByteWithTimeout(devicePort, (short) 10);
   GetStringWait(stringin, (short) 2000);
   if (strstr(stringin, ">") != NULL) r=true;
   else r=false;
